requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib
argon2-cffi
python-jose